crontab-cambridge: Change the days when we run a given distro-debian suite
The weekly CD images which are used by the snapshot flight are
generated Sunday-Monday, so running that on a Saturday as we have been
doing ensures that it will take at least two iterations/weeks to get
any issues fixed.

Also the current ordering of the existing releases made it hard to
decide where to insert a new release (e.g. Stretch).

So reorder as:
 - Run the Sid daily run on a Monday
 - Run the Snapshot run on a Tuesday (to pick up the weekly builds
   from Monday)
 - Run Squeeze on Wednesday and continue with newer releases
   chronologically from there.

New releases can then be added at the end (wrapping the days).

Also add some blank lines to aid clarity.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>