mg-show-flight-runvars: Decorate synth runvar names with ~
Make mg-show-flight-runvars -a append ~ to the names of synth runvars.
(This is consistent with the new syntax in cs-job-create.)

We do this by editing $row[1] (and $colws[1]) so we can avoid
disturbing the general column format calculation and printing.

We switch to fetchrow_array rather than fetchrow_arrayref.  This is
clearer and also avoids having to copy $row (because the value in the
DB $row from fetchrow_hashref would be readonly).

We have to check for $synth eq 'f' as well as $synth being boolean
false, because SQLite's typeless nature (or, to put it another way,
DBD::SQLite's failure to look at the schema) means that a boolean
field's value of 'f' or 't' is simply returned as a string to Perl.
But of course "f" is trueish in Perl.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>
---
v2: Use fetchrow_array instead.
    Do not mistakenly drop the $synthcond assignment (!)