Debian i386 HVM tests: Increase installation timeout
The Debian i386 image boots a 32-bit non-PAE kernel which therefore
cannot have any PV drivers on our 64-bit hypervisors.  This makes it a
bit slow: in my test on a machine under my desk it took 1400s out of
the allowed 2000s.

With this change the timeout is 3000s instead.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>