Executive HTML output: Use #888888 (grey) for queued jobs
Either the flight hasn't started yet, or the job is blocked waiting
for other jobs to finish.  In any case their state is not very
interesting.

Most usefully this change visually distinguishes, in the plan summary,
jobs which are waiting for prior jobs to finish, from ones which have
entered the planning queue.

Also replace a $f->{status} with $status, which is less confusing.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>
---
v2: Get test right.
    Update commit message.