ts-logs-capture: Actually do hard host reboot sometimes
The logic in try_fetch_logs for setting $ok was wrong.  $ok would be
set if we reached the end of any outer (pattern) loop iteration.  If
the host is actually dead all the pattern expansions would fail, but
some of the patterns are literals and do not need expansion.  The
inner (logfile) loop would say `next' if the logfile fetch failed, but
that just goes onto the next logfile.  So this code would always set
$ok.

Instead, set $ok to 1 when we successfuly fetch any logfile or
successfully expanded any pattern (even if it didn't match any files).

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>