Debian: Enable interpolation in uboot_scr_load_dtb here doc
By switching <<'END' to <<END. A future patch is going to want to put
a variable here which requires interpretation by the Perl.

Unfortunately this means lots of extra backslashes to escape things
such that they pass through Perl and Shell and end up as ${foo} in the
resulting u-boot script.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>