ts-kernel-build: Explicitly enable CONFIG_CGROUPS
This is not enabled in the arm defconfig (it is in x86) and without it
a Debian Jessie userspace with systemd fails to boot with:

systemd[1]: Failed to mount tmpfs at /sys/fs/cgroup: No such file or directory

While we arrange to use sysvinit in the host and in d-i installed
guests we do not do so for xen-create-image guests.

Arranging to install sysvinit-core during xen-create-image looks to be
non-trivial. I hope it won't be necessary.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>