Debian: Arrange to drop "quiet" from command line of installed system
By editing /etc/default/grub in a late command iff it exists.

This will affect ts-debian-{hvm,di}-install as well as
ts-host-install and hence affect guests as well as hosts.

The overall effect is that we will log more upon guest boot as well as
on the initial host boot.

Note that for hosts we also strip the "quiet" when updating grub as
part of ts-xen-install, this code remains (although it won't do so
much now).

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>