ap-fetch-*: Support $AP_FETCH_PLACEHOLDERS envvar which outputs a placeholder
And use this in standalone-generate-dump-flight-runvars. In general I
don't think we are interested in the specific revision_* runvars when
using this tool but when it matters this new behaviour can be avoided
by setting AP_FETCH_PLACEHOLDERS=n.

This is quicker even than using memoisation on the ap-fetch
invocations and produces output like:

libvirt  build-amd64  revision_xen  ap-fetch-version-baseline:xen-unstable

This is useful when doing comparisons of before and after changes to
e.g. make-flight since they do not pickup noise if a something/someone
does a push in the middle.

The memoisation bits of standalone-generate-dump-flight-runvars are
disabled if AP_FETCH_PLACEHOLDERS=y.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>