ms-queuedaemon: Do not spin if client input is delayed/truncated
chan-read-data would spin if `read' returns early because of
nonblocking mode.

Check whether the return value is the empty string (which can only
happen on eof or nonblocking lack of data, and we checked eof just
before), and if so, simply return.  The fileevent remains set up so we
will be called again when more data arrives.

(Deployment/testing note: this change is currently live in Cambridge,
as I cowboyed it directly into ~osstest/daemons-testing.git, on
observing this misbehaviour.)

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>