Nested hosts: Use hostnamepath() in create_webfile
create_webfile needs a pathname in the shared public-html directory.
These paths need to be (a) stable (b) unique across all running jobs.
We achieve this by basing the filenames on the hostname and (for a
guest) the guest name.

But for an L2 guest we need to include the physical host name too,
because the L1 `host' is not unique.

Fix this by using hostnamepath(), replacing the open-coded single
iteration.

Reported-by: Ian Campbell <ian.campbell@citrix.com>
CC: Robert Ho <robert.hu@intel.com>
Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>