Log capture pathnames: Include host name (and nested L0 name)
Rather than using $gho->{Name} (and, sometimes, separately
$ho->{Name}), use hostnamepath().

This means that the hostname is always included in the standard way.
The filename structure for xenctx and console snapshots changes
slightly.

In two cases this means that the hostname is now included where it
wasn't previously.  That is helpful if due to some insanity the same
guest is present on more than one host.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>